Break out pkg for server code
