+++ pkg/transport/transport.go
@@ -1,0 +1,11 @@
+package transport
+
+import "net/http"
+
+func CopyHeaders(destination http.Header, source *http.Header) {
+	for k, v := range *source {
+		vClone := make([]string, len(v))
+		copy(vClone, v)
+		(destination)[k] = vClone
+	}
+}
